# task11/export/zrhev5wrua_oas_v3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "task11_api",
    "version": "2024-08-12T11:40:13Z"
  },
  "servers": [
    {
      "url": "https://zrhev5wrua.execute-api.eu-central-1.amazonaws.com/{basePath}",
      "variables": {
        "basePath": {
          "default": "api"
        }
      }
    }
  ],
  "paths": {
    "/reservations": {
      "get": {
        "summary":"Get Reservation",
        "description":"This endpoint is used to get all reservations details.",
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationsGETResponse"
                }
              }
            }
          },
          "400": {
            "description": "400 Bad Request",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_templates",
          "type": "aws_proxy"
        }
      },
      "post": {
        "summary":"Add new reservation",
        "description":"This endpoint is used to add reservations details.",
        "requestBody":{
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationsPOSTRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationsPOSTRequest"
                }
              }
            }
          },
          "400": {
            "description": "400 Bad Request",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_templates",
          "type": "aws_proxy"
        }
      },
      "options": {
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Methods": "'*'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "when_no_match",
          "type": "mock"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReservationsPOSTRequest": {
        "title": "Reservations POST Request",
        "type": "object",
        "properties": {
          "tableNumber": {
            "type": "integer"
          },
          "clientName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "slotTimeStart": {
            "type": "string"
          },
          "slotTimeEnd": {
            "type": "string"
          }
        }
      },
      "ReservationsGETResponse": {
        "title": "Reservations GET Response",
        "type": "object",
        "properties": {
          "tableNumber": {
            "type": "integer",
            "example": 100
          },
          "clientName": {
            "type": "string",
            "example": "Test Client"
          },
          "phoneNumber": {
            "type": "string",
            "example": "9876543210"
          },
          "date": {
            "type": "string",
            "example": "2024-06-20, yyyy-MM-dd format"
          },
          "slotTimeStart": {
            "type": "string",
            "example": "13:00"
          },
          "slotTimeEnd": {
            "type": "string",
            "example": "16:00"
          }
        }
      }
    },
    "securitySchemes": {
      "authorizer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "x-amazon-apigateway-authtype": "cognito_user_pools",
        "x-amazon-apigateway-authorizer": {
          "x-syndicate-cognito-userpool-names": ["simple-booking-userpool"],
          "type": "cognito_user_pools"
        }
      }
    }
  }
}